requirements.txt
Flask
flask-sqlalchemy
email-validator
flask-msearch
flask-login
flask-bcrypt
flask-wtf
requests
pillow
gunicorn
psycopg2